requirements.txt
fastapi==0.116.1
httpx==0.28.1
matrix-nio==0.25.2
pydantic==2.11.7
pydantic-settings==2.10.1
PyYAML==6.0.2
uvicorn[standard]==0.35.0
